moving ltq executor to test code
